Switch to Jena uber-JAR when running from Eclipse
When running from an OSGi environment, we need to have all the Jena libraries
in the same classloader. Otherwise, Jena will not be able to find all its
components through the Java ServiceLoader.

This means that we have to package everything we need from Jena into a single
uber-JAR. The new "jena-uberjar" folder has a Maven build that will produce this
locally. Developers will need to install this uber-JAR into their local Maven
repositories before they can work on the code.