Return single values if OWL indicates the property's maximum cardinality is 1 (#22)
This commit uses the OWL reasoner in Jena to expand the property graph, and uses the Jena Ontology API to detect if the property being accessed has a maxCardinality restriction. It will limit the number of returned elements to that maxCardinality, and if it is 1, it will either return that value or `null` if no value exists.

This is in order to follow the conventions in the OWL ontologies produced by the MOF2RDF mapping. This specific use of restrictions (and other MOF2RDF conventions) will be moved in the future to a MOF2RDF-specific subclass of RDFModel, leaving RDFModel as a "pure RDF" model that acts as an Epsilon wrapper over Apache Jena.

---------

Co-authored-by: Antonio Garcia-Dominguez <a.garcia-dominguez@york.ac.uk>