Refactoring getProperty() added getCollectionOfPropertyValues()
Resource is now public to enable extensions of RDFResource access.

listPropertyValues doesn't need to be public.

Removed the MaxCardinality checking and value return processing.

Reordered the methods involved in the getProperty process, to try and
improve readability. This entire chain of methods could do with a
refactor at some point.