Separate MOF2RDF specifics to its own model subclass
There may be users who want to use property values exactly as they come from the RDF graph, without any postprocessing from the model driver. Rather than overcomplicate things with toggles and so on, it's best to have a core class (RDFModel) which stays as close to Jena as possible, and then have subclasses that implement some postprocessing of the Jena API results according to certain conventions (e.g. those of the MOF2RDF spec).

---------

Co-authored-by: Antonio Garcia-Dominguez <a.garcia-dominguez@york.ac.uk>