Update RDF statements when a single-valued EAttribute is changed (#59)
This commit adds support for updating the RDF statements when a single-valued attribute is changed in an EObject. This is meant to be the start of a series of changes to support the creation and modification of EMF objects on top of an existing RDF graph, without losing other information that may not be part of the EMF metamodel.

---------

Co-authored-by: Antonio Garcia-Dominguez <a.garcia-dominguez@york.ac.uk>