Add support for changes in multi-valued EAttributes (#61)
This commit adds sync logic for handling multi-valued EAttribute fields,
which can be represented as RDF bags, sequences, or lists. Two new
suites of change-based equivalence tests have been added, which
both inherit from a common superclass to make it easier to support
more change-based equivalence tests in the future.

The commit also fixes updates of single-value attributes with dates,
by falling back on literal-to-object conversion.

xsd:dateTime is tricky in that the same instant in time can be represented
through multiple strings, based on timezone. For this reason, we need to
have a fallback when doing single attribute updates:

* First, we try object-to-literal conversion as usual.
* If we can't find the old statement for it, we go through all
  the statements for that resource+property, deserialise them,
  and do an actual comparison.
* If that still doesn't work, we print an error message.

I have changed some of the test data to use a timezone that we're unlikely
to use ourselves (UTC+8), to reproduce this issue.

---------

Co-authored-by: Antonio García-Domínguez <a.garcia-dominguez@york.ac.uk>