Replace XMI with one that does not use xmi:id
Given an EObject, the DefaultMatchEngine in EMF Compare will pick a
matching algorithm based on whether it has an ID or not. If it has
an ID (e.g. an `xmi:id`), it will use ID-based matching. Otherwise,
it will use structural matching.

The test `.xmi` we had for single EReferences was using XMI IDs,
whereas the EObjects from our `.rdfres` files do not have any such
IDs. This was causing unnecessary differences to be detected by EMF
Compare, due to the objects not being correctly matched to each other
from the start.