Relax deserialiser: log classes not found in EPackage instead of aborting deserialisation
We noted that in some cases, the EPackage created to represent the types of
interest may not have every class in a vocabulary. From an open-world point
of view, in these scenarios it would make more sense to log the missing type
and continue loading, rather than completely aborting the load. It may be
that the user is really only interested in some of the concepts in that
vocabulary.