Add support for EObject creation and deletion
This commit implements the creation and deletion of EObjects in the resource,
including the recursive creation and deletion of their descendants in the
containment hierarchy.

The documentation has been updated to describe the required tweaks to the .rdfres
file when using the resource to create new RDF files. The user should either set a
default namespace IRI for the new instances (which use EMF-generated UUIDs),
whether through the .rdfres, or by setting the namespace IRI of the empty prefix
(e.g. `PREFIX : <http://foo/bar/> .`).

If none of the above are set, the resource will fall back to Jena's default behaviour,
which is to use the absolute IRI of the RDF model (e.g. `file://a/b/c/file.ttl#`).

---------

Co-authored-by: Antonio Garcia-Dominguez <a.garcia-dominguez@york.ac.uk>