Fix synchronisation of model roots (fixes #73)
The adapter now handles ADD, ADD_MANY, REMOVE, and REMOVE_MANY at the resource level. There are tests to check for this as well.

The adapter propagation test had to be rewritten, as we need to have a Jena graph set up in place before we can make any changes to the model (otherwise, we have nothing to synchronise!).

---------

Co-authored-by: Owen Reynolds <owen.reynolds@york.ac.uk>