Add load equivalence test for 'most specific type' scenario and fix findMostSpecificEClasses (fixes #78)
We didn't have a test for the scenario where the same resource is associated to multiple EClasses, where some of the EClasses are supertype of the others. This commit adds such a test, and fixes an issue caught by the test which would trigger if we ran into the more general type *after* the most concrete type.

The commit also changes RDFResourceConfiguration to use lists instead of sets, to ensure that we always honor the order in which the data models and schema models are listed in the `.rdfres`.