rdf.emf: make org.eclipse.core dependencies optional
This dependency was due to our use of FileLocator to resolve platform:/ and resource:/ URLs into file:/ URLs when running the resource from Eclipse. The dependency creates unnecessary problems when using the resource from plain Java, due to how JAR signer information has been historically produced in the org.eclipse.core artifacts.

This commit changes the code to use FileLocator strictly via reflection: if it is not available (e.g. when using the resource from plain Java), it will simply be skipped over (we will use URLs as-is).