Add support for modifying the resource without loading it (#93)
This PR adds support for modifying the resource without having loaded it. When that happens, a union of in-memory data models will be set up based on the configuration for the resource at that time. If no data models are defined, a .ttl data model will be added based on the URI to the .rdfres (which will need to be a file: or platform: URI).

In such a case, the configuration will also be saved when the resource is saved (not just the data models).